requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-multipart==0.0.9

torch==2.1.2
transformers==4.40.0
accelerate==0.29.2
sentence-transformers==2.6.1

langchain==0.1.13
langchain-community==0.0.30
langchain-core==0.1.44
langchain-huggingface==0.0.5

faiss-gpu==1.7.2
Pillow==10.2.0
tqdm==4.66.4
